@@ -1,28 +1,56 @@
-buildscript {
-    repositories {
-        mavenCentral()
-    }
-    dependencies {
-        classpath 'com.android.tools.build:gradle:0.9.+'
-    }
-}
+/**
+ * The first line in the build configuration applies the Android plugin for
+ * Gradle to this build and makes the android block available to specify
+ * Android-specific build options.
+ */
 
+apply plugin: 'com.android.application'
+apply plugin: 'com.google.gms.google-services'
 
-repositories {
-    mavenCentral()
-}
-
-apply plugin: 'android'
+/**
+ * The dependencies block in the module-level build configuration file
+ * specifies dependencies required to build only the module itself.
+ * To learn more, go to Add build dependencies.
+ */
 
 dependencies {
-    compile 'com.google.code.gson:gson:2.2.4'
-    compile 'com.squareup.picasso:picasso:2.1.1'
-    compile 'com.android.support:support-v4:13.0.0', 'com.google.android.gms:play-services:3.1.36'
+    implementation platform('com.google.firebase:firebase-bom:28.3.0')
+    implementation 'com.google.firebase:firebase-messaging'
+    implementation 'com.google.firebase:firebase-analytics'
+
+    implementation 'com.google.code.gson:gson:2.8.7'
+    implementation 'com.squareup.picasso:picasso:2.1.1'
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
+    implementation 'com.google.android.gms:play-services-location:18.0.0'
+    implementation 'com.google.android.gms:play-services-maps:17.0.1'
+
+    //implementation project(":lib")
+    implementation 'androidx.appcompat:appcompat:1.3.1'
+    implementation fileTree(dir: 'libs', include: ['*.jar'])
+    implementation 'com.google.android.material:material:1.3.0'
 }
 
 android {
-    compileSdkVersion 18
-    buildToolsVersion "19.0.3"
+
+    useLibrary 'org.apache.http.legacy'
+
+    /**
+     * compileSdkVersion specifies the Android API level Gradle should use to
+     * compile your app. This means your app can use the API features included in
+     * this API level and lower.
+     */
+
+    compileSdkVersion 30
+    /**
+     * buildToolsVersion specifies the version of the SDK build tools, command-line
+     * utilities, and compiler that Gradle should use to build your app. You need to
+     * download the build tools using the SDK Manager.
+     *
+     * This property is optional because the plugin uses a recommended version of
+     * the build tools by default.
+     */
+
+    buildToolsVersion "30.0.2"
 
     signingConfigs {
         debug {
