@@ -1,7 +1,7 @@
 package com.profete162.WebcamWallonnes.Utils;
 
 import android.location.Location;
-import android.support.v4.app.Fragment;
+import androidx.fragment.app.Fragment;
 
 import com.profete162.WebcamWallonnes.CamFragment;
 
