@@ -15,8 +15,11 @@ import android.view.ViewGroup;
 import android.view.WindowManager;
 import android.widget.ImageView;
 
+import androidx.annotation.NonNull;
+
 import com.google.android.gms.maps.CameraUpdateFactory;
 import com.google.android.gms.maps.GoogleMap;
+import com.google.android.gms.maps.OnMapReadyCallback;
 import com.google.android.gms.maps.SupportMapFragment;
 import com.google.android.gms.maps.model.BitmapDescriptorFactory;
 import com.google.android.gms.maps.model.LatLng;
@@ -36,7 +39,7 @@ import java.util.ArrayList;
 import java.util.Collections;
 import java.util.Comparator;
 
-public class ZeMapFragment extends SupportMapFragment {
+public class ZeMapFragment extends SupportMapFragment implements OnMapReadyCallback {
 
     Location location;
     GoogleMap mMap;
@@ -49,28 +52,9 @@ public class ZeMapFragment extends SupportMapFragment {
         setHasOptionsMenu(true);
 
         // mMapFragment = MapFragment.newInstance();
-        mMap = this.getMap();
-        if (mMap != null) {
+        getParentFragment().getActivity().setProgressBarIndeterminateVisibility(false);
 
-            location = ((DrawerActivity) this.getActivity()).loc;
-
-            CamRequest camTask = new CamRequest();
-            camTask.execute("CVE");
-            mMap.setInfoWindowAdapter(new PopupTrafficAdapter(getParentFragment().getActivity().getLayoutInflater()));
-
-
-            preferences = PreferenceManager.getDefaultSharedPreferences(getParentFragment().getActivity());
-
-            mMap.moveCamera(CameraUpdateFactory.newLatLngZoom(
-                    new LatLng(preferences.getFloat("lat", 50.5f),
-                            preferences.getFloat("lng", 4)), preferences
-                    .getFloat("zoom", ZOOM)));
-
-
-        }else
-            getParentFragment().getActivity().setProgressBarIndeterminateVisibility(false);
-
-
+        this.getMapAsync(this);
     }
 
     public void doStuff() {
@@ -127,6 +111,26 @@ public class ZeMapFragment extends SupportMapFragment {
 
     }
 
+    @Override
+    public void onMapReady(@NonNull GoogleMap googleMap) {
+        this.mMap = googleMap;
+
+        location = ((DrawerActivity) this.getActivity()).loc;
+
+        CamRequest camTask = new CamRequest();
+        camTask.execute("CVE");
+        mMap.setInfoWindowAdapter(new PopupTrafficAdapter(getParentFragment().getActivity().getLayoutInflater()));
+
+
+        preferences = PreferenceManager.getDefaultSharedPreferences(getParentFragment().getActivity());
+
+        mMap.moveCamera(CameraUpdateFactory.newLatLngZoom(
+                new LatLng(preferences.getFloat("lat", 50.5f),
+                        preferences.getFloat("lng", 4)), preferences
+                        .getFloat("zoom", ZOOM)));
+
+    }
+
     public class APITrafficRequest extends AsyncTask<String, Void, TrafficFragment.ApiResponse> {
 
         @Override
@@ -141,7 +145,7 @@ public class ZeMapFragment extends SupportMapFragment {
 
 
                 try {
-                    String url = "http://data.beroads.com/IWay/TrafficEvent/" + getString(R.string.lan) + "/all.json?format=json&from="
+                    String url = "https://data.beroads.com/IWay/TrafficEvent/" + getString(R.string.lan) + "/all.json?format=json&from="
                             + location.getLatitude()
                             + ","
                             + location.getLongitude()
@@ -198,7 +202,7 @@ public class ZeMapFragment extends SupportMapFragment {
         protected RadarFragment.ApiResponse doInBackground(String... params) {
 
             if (location != null) {
-                String url = "http://data.beroads.com/IWay/Radar.json?format=json&from="
+                String url = "https://data.beroads.com/IWay/Radar.json?format=json&from="
                         + location.getLatitude()
                         + ","
                         + location.getLongitude() + "&max=30";
@@ -288,7 +292,7 @@ public class ZeMapFragment extends SupportMapFragment {
                                     Intent streetViewIntent = new Intent(Intent.ACTION_VIEW, streetViewUri);
                                     startActivity(streetViewIntent);
                                 } catch (Exception e) {
-                                    startActivity(new Intent(Intent.ACTION_VIEW, Uri.parse("http://play.google.com/store/apps/details?id=com.google.android.street")));
+                                    startActivity(new Intent(Intent.ACTION_VIEW, Uri.parse("https://play.google.com/store/apps/details?id=com.google.android.street")));
                                 }
                                 break;
                             default:
